Partially port loader
loader has been partially ported. To finish porting loader, support for
each region beyond USA and KOR must be added, and the size of
the SystemHeap must be determined, and the SystemHeap must be correctly
expanded (both of which have yet to be done). See `loader.cpp` for more
information.

Beyond this, loader still does not work in this version because of an
issue in a Petari header. I am committing what I have now before I go
fix the issue.